rearanged wiki publish components to handle enough space for publish logic
