+++ src/components/CreateWiki/CreateWikiTopBar/index.tsx
@@ -2,7 +2,7 @@ import { HStack } from '@chakra-ui/react'
 import React from 'react'
 import { WikiTitleInput } from './WikiTitleInput'
 import WikiScoreIndicator from '../WikiScoreIndicator'
-import { WikiPublishButton } from './WikiPublishButton'
+import { WikiPublishButton } from './WikiPublish/WikiPublishButton'
 
 export const CreateWikiTopBar = () => {
   return (
